+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.82.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.82.0)
+_July 16, 2025_
+
 ## StreamChat
 ### 🐞 Fixed
 - Fix channel unread count cleared when a thread is marked as read [#3710](https://github.com/GetStream/stream-chat-swift/pull/3710)